pyproject.toml
[tool.mypy]
mypy_path = "src/"
ignore_missing_imports = true


[tool.pytest.ini_options]
addopts = "--rootdir ../tests/ --ignore=__pypackages__"


[project]
name = "app"
version = "0.0.1"
description = ""
authors = [
    {name = "Leandro de Souza", email = "leandrodesouzadev@gmail.com"},
]
dependencies = [
    "fastapi[all]==0.74.1",
    "tortoise-orm[asyncpg]==0.18.1",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "email-validator>=1.1.3",
    "virtualenv==16.7.7",
    "python-dotenv>=0.19.2",
    "colorama>=0.4.4",
]
requires-python = ">=3.8"
license = {text = "MIT"}

[project.urls]
Homepage = ""

[project.optional-dependencies]
[tool]
[tool.pdm]
[tool.pdm.dev-dependencies]
dev = [
    "pytest>=7.0.1",
    "black>=22.1.0",
    "pytest-mock>=3.7.0",
    "tox-pdm==0.3.3",
    "mypy>=0.931",
    "tox>=3.24.5",
]

[build-system]
requires = ["pdm-pep517"]
build-backend = "pdm.pep517.api"
